* commit '6b1f9d4198358f2a586478daa02e8a9a18aec8c6':
  TFC-43: Refactored time to seconds, cleaned up enemy, added fps counter

+++ src/EnemyManager.h
@@ -19,7 +19,7 @@ public:
 private:
     const Map &map;
 public:
-    void update(long long int delta);
+    void update(float delta);
 
     explicit EnemyManager(const Map &map);
